Update patch set 1
Patch Set 1:

I unfortunately don't have any great suggestions here.  In the past we've patched the project to deal with these Android-isms, but I'm not suggesting that it's a great idea either.

I'm honestly pretty okay with this change as is.  If this is a bigger, generic problem, we could likely make a change to ueventd to automatically create a symlink to the typical linux nodes.

+sspatil@ who is likely interested in this conversation too.

Patch-set: 1
Reviewer: Gerrit User 1108565 <1108565@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>