Update patch set 1
Patch Set 1:

> Does that actually happen, though? It seems like if you want to do that, you can just use a static library for your internal shared code?

Not yet, just because the use of version_script is not popular or enforced I guess. I am not sure if I understand what you mean by 'use a static library for your internal shared code', but if it means for example statically linking libadbconnection_client from the other files in the same APEX, while requiring entities outside of the APEX to dynamically linking to it, then I think it's not efficient because libadbdconnection_client code is duplicated (one copy in the *.so file itself, and the other copy in the other files in the same APEX). That might also cause ODR violation.

Patch-set: 1