Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> It does seem wrong that it uses "symlink" over /dev/binder. On kernels prior to this patch, that will overwrite the device node with (probably) a dead link, won't it?

No, the symlink command will fail harmlessly if /dev/binder exists. It seems like the issue here is having these commands present in early-init right after ueventd starts which leads to a race between the creation of binder devices and the execution of the symlink command. I was able to verify that moving them to init instead(when ueventd is guaranteed to have completed device creation) fixes boot on one of the broken devices.

Patch-set: 1