Update patch set 1
Patch Set 1:

> that all sounds like a feature to me, personally.

Well, yes, it would be an (annoying) feature, I think.
 
> as an alternative though we should probably at least have a blacklist for stuff like libc++ and libbase and anything used in multiple modules and/or used in a module and on the system image, so you can never re-export any of their symbols. that's never legit, right?

Right, that's not legit.. for a library that is shared across different linker namespaces like libadbdconnection_client.so. It should be fine if the lib wasn't shared from adbd to art apex.

I see that Anton filed a bug for this. Let's discuss more there.

Patch-set: 1