Update patch set 1
Patch Set 1:

Did this actually fix the test though?  I don't see many new test runs on the test history link in the bug.  I was hoping to see them all turn green after submitting this.

Otherwise, do we have test runs with the dump of the unexpected log tags and comments from the other CL that we submitted to try to understand what's going on here?

Patch-set: 1