Update patch set 1
Patch Set 1:

> Patch Set 1: Presubmit-Ready+1
> 
> Hi all,
> 
> This is an RFC: do you think we should create individual SDKs for platform related things? Or should we create only one SDK? If only one SDK which git repo for expressing the build file do you think would be its natural place?

Ideally, I think we should have one SDK for the whole platform. Whether that is practical or not I don't know. If we had to have a couple of SDKs for the platform that would be ok. Having one per repository contributing to the platform might be ok too. Having one per library would probably not be ok.

Where to put it? I don't know. My gut feel is to have one per repository that contributes to the platform for now simply to avoid cross repository dependencies which may come back to bite us in future modularization work.

Patch-set: 1
Reviewer: Gerrit User 1039338 <1039338@85c56323-6fa9-3386-8a01-6480fb634889>