Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > Has the kernel patch been taken yet?
> 
> No. Hopefully Catalin will take a look this week.
> 
> > It seems too early to make this change until it is in a kernel.
> 
> Hmm, that would make it hard to land the debuggerd changes to support MTE on top of this one.

yeah, i think there's a chicken and egg situation here. i'm normally a stickler for "get it upstreamed before you even talk to us", but MTE seems like it's too big and complicated and we're going to be doing so much of the early work that i'd rather we had as much early experience as possible.

> What do you think about putting this behind ANDROID_EXPERIMENTAL_MTE? We already have code to support unlanded kernel changes behind that flag.

is there any benefit to that? in general if we can have something building and running for everyone, i prefer that over having to have special builds.

Patch-set: 1
Reviewer: Gerrit User 1003224 <1003224@85c56323-6fa9-3386-8a01-6480fb634889>