Update patch set 2
Patch Set 2:

(2 comments)

What's the perf impact when operating on a non-sorted symbol table? Also, what's the RSS impact (assuming that one reason to make this change)?

I ask because I think almost all symbol tables that we look at are not sorted, so this is only going to speed up the art ones. The worst case is that there is no function for the given pc. I didn't look closely, but in that case I think you would still search the entire symbol table. And this occurs on many unwinds because not every function is in the symbol table.

Patch-set: 2
Reviewer: Gerrit User 1019050 <1019050@85c56323-6fa9-3386-8a01-6480fb634889>