Update patch set 1
Patch Set 1: Code-Review-2

Sorry, I appreciate having this and will look into these and probably take the bug, but we can't submit this as is. If someone blindly changes the service name and rewrites all of the rules involving it, they may break the GSI in a way that we won't detect until testing time.

Patch-set: 1
Reviewer: Gerrit User 1120458 <1120458@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2