Update patch set 6
Patch Set 6:

(2 comments)

I ran a custom benchmark, and the memory savings are impressive, everything practically goes down to not detectable increases in RSS. It does look like the cost of lookup is about 3.6% worse when doing a random lookup and about 6.8% worse when trying to find a symbol that doesn't exist. Those are both for the first time doing the lookup in a shared so that is not sorted. On the bright side, it is significantly faster (about 21% better) for the sorted oat files doing a lookup where the symbol is found.

The example shared so file and oat file I chose are not necessarily going to be the examples of all files, but I think they are close enough to demonstrate this is a good change.

Patch-set: 6
Reviewer: Gerrit User 1019050 <1019050@85c56323-6fa9-3386-8a01-6480fb634889>