Update patch set 7
Patch Set 7:

I reran my benchmark, which I'll upload right after this.

Everything is pretty good, but there is one outlier in the data. When resolving multiple symbols from a non-sorted symbol table, the time is about 150% worse that the normal mode. I would guess that this is because of the caching that the previous code does versus how each search is probably not caching the same data so each search is worst case.

Patch-set: 7
Reviewer: Gerrit User 1019050 <1019050@85c56323-6fa9-3386-8a01-6480fb634889>