Update patch set 8
Patch Set 8: Code-Review+2

It looks like this every so slightly slows down unwinding. In the normal case, it looks like a 2% decrease. Given the memory improvements, I think this is acceptable. For heapprofd, I think it should even out as more unwinds go through the same symbol. The worst case would be a single unwind not reusing any symbols. Even debuggerd does multiple thread unwinds and would likely be less affected.

I don't like the second memcpy of the symbol data, and their is probably a way to do something a bit better. There is a way to tell the compiler that there is an unaligned access so it won't create the code that will cause an unaligned access. I'll see if I can dig that up, but no reason to hold this change for that.

Patch-set: 8
Reviewer: Gerrit User 1019050 <1019050@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2