Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > how much difference does this make?
> > 
> > can we get a similar effect by making some of the member functions free functions instead and just letting the linker gc them?
> 
> For coverage purposes, no, because right now, functions that get gced by the linker still count against you in coverage (b/151634315)

how much difference does this make to the percentage? i feel like -- and correct me if i'm wrong -- unlike the other changes in this stack, this is one that you wouldn't otherwise make. the others are all improvements, where as this feels like it makes the code a bit worse to work around the tool limitation?

should we wait and see whether pirama can just get this automated in time? we can always do this at the last minute if he can't...

Patch-set: 2
Reviewer: Gerrit User 1003224 <1003224@85c56323-6fa9-3386-8a01-6480fb634889>