Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > are we adding METADATA to *all* projects, including the ones that only contain code we wrote? (i've seen you removing MODULE_* files, but not adding METADATA to those same projects...)
> 
> I am adding METADATA to all projects that contain 3p code. Per the go/android3p definition, BSD is 3p code. In platform/external, I assume any Apache2-only code is there to be vendored in so I treat it as 3p regardless.

oh, is this for libpixelflinger (which has also been deleted in internal master)?

should we merge this here but then remove it again in rvc-dev-plus-aosp? (i don't know whether the presence of this METADATA file makes it easier to add _new_ non-Apache-2.0? that's something we definitely don't want --- we're clean internally now, and would like to stay that way.)

Patch-set: 3
Reviewer: Gerrit User 1003224 <1003224@85c56323-6fa9-3386-8a01-6480fb634889>