Update patch set 4
Patch Set 4: Code-Review+2

Note that boot images are usually flashed in the bootloader, not fastbootd (flashstation and host fastboot only issue "reboot fastboot" before flashing dynamic partitions).

If that's going to be a problem, it might be better to do this logic in the host tool.

Patch-set: 4
Reviewer: Gerrit User 1290458 <1290458@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2