Update patch set 1
Patch Set 1:

(1 comment)

Thanks for finding and pushing this.  I wonder if there's any chance this is related to b/150863651?  This issue seems like it's not recoverable, whereas that issue is, but I'd love for this patch to fix that bug.

Overall: We really need to have a longer term solution for this USB code and init.  This code was never reasonable to include in init and the fact that it is giving us issues should give us renewed effort to move it to a dedicated USB service.

Patch-set: 1
Reviewer: Gerrit User 1042384 <1042384@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1373349 <1373349@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>