Update patch set 6
Patch Set 6:

For what it's worth, I'm a bit confused at the usefulness of this.  

The example .05s delay is going to wait virtually no time for the file to be present.  If there is variance greater than .05s in the time that the file is going to be present, then such a small wait will finish before the file was present even on devices where the file is eventually present in some percentage of boots.

Otherwise, if .05s is enough time to wait for the file if it will be eventually present 100% of the time, I assume that the file would have been present before this wait, since a 50ms window is very tight compared to the whole boot process.

Are there better mechanisms than having init wait for the file that we can use?  What is the larger intention for this use case?

Patch-set: 6
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>