Update patch set 4
Patch Set 4: Code-Review-2

(22 comments)

I don't think any of these fuzzers are actually doing much that is useful. I think you could restructure this to be something better. For example, here is some kinds of good fuzzing:

- Randomize the data that is interpreted as unwind information. This would probably be good at finding weird corner cases that might provoke crashes.
- Randomize the stack data along with the unwind information. Use different size initial stack data, and random data which could provoke some kind of crashes.
- Randomize jit unwind data.
- Randomize dex file data.
- Randomize the initial register data.

Doing the above would require doing a bit more setup to create a fake kind of elf, and a fake set of stack data that can be passed to an unwind. You would also have to create the fake jit data and also the fake dex data so that you are actually using a real dex with some part of it being randomized.

I think this would be better as a single fuzzer just for Unwind and maybe some variations on the above if you really want to separate them out, maybe a dex file fuzzer, a jit fuzzer, and so on. Given the way the code works, I think that might be the best, because varying everything is not as useful because each piece is really done in a way that it's distinct from the others. You either use the uwnind data from the elf, or the jit data. The dex file is a bit different, but it's also orthogonal to the other things, so having a fuzzer just for it would probably be useful.

Patch-set: 4
Reviewer: Gerrit User 1019050 <1019050@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2