Update patch set 4
Patch Set 4:

Just to add some additional context here, the general structure of most of the fuzzers revolves around a vector of lambda functions to modify state. The fuzzer defines the index of the function to call, and passes in whatever data we need. In the example of unwinder, it passes a pointer to the fuzzeddataprovider and the unwinder, then generates any additional needed variables within each lambda function. This allows the fuzzer to control the order in which functions are called, as well as the contents of the functions themselves.

Patch-set: 4