Update patch set 6
Patch Set 6:

(1 comment)

Glad to see that we can start ueventd in the default mount namespace ahead of time.  

I have a handful of questions though:
This means that the bootstrap libc/linker/linker config are available in the default mount namespace?  I was figuring that they were not available and that would be the blocker here.  Is that true even after we handle the APEX mounts?

Does this mean that if ueventd restarts after APEXs are mounted that it uses the APEX version of libc, since it's in the default mount namespace?

Is there a way to indicate to the linker "I'm in the default mount namespace but I really want the bootstrap version of libc"

If all of the above is possible, why do we need the two mount namespaces?  I'm sure there's a reason but I'm just curious given the above.

Patch-set: 6