Update patch set 5
Patch Set 5:

Hi Elliott, It seems like the blocking bug b/128018586 still need some time to be fixed. In the meanwhile I'd like to submit this CL first to keep the build clean. We should definitely come back and revisit this when b/128018586 is unblocked. (or raise the severity of it if you think this deserves more attention?)

Patch-set: 5