Update patch set 2
Patch Set 2:

I really hate vector::erase...

The proper fix is to rewrite it as `vector->erase(std::remove_if(...), vector->end())`, which is a no-op if the iterator returned by remove_if is vector->end(). As written here, if remove_if finds multiple entries, we'll move all of them to the end, but only delete the first one of them.

(I'll grep through the codebase tomorrow to see if we've introduced any other bugs of this form since the last time I've seen this bug. I've been CCed on at least 3 other instances of this that I can find via gerrit, and I'm 90% sure there's two others that I can't find)

Patch-set: 2